[STYLE] Move geojson converter to utils
