+++ app/api/src/utils.py
@@ -8,6 +8,8 @@ import random
 import re
 import shutil
 import string
+import subprocess
+import tempfile
 import time
 import uuid
 import zipfile
@@ -38,7 +40,6 @@ from starlette import status
 from starlette.responses import Response
 
 from src.core.config import settings
-from src.core.heatmap.heatmap_core import convert_geojson_to_others_ogr2ogr
 from src.resources.enums import MaxUploadFileSize, MimeTypes
 
 
@@ -1050,6 +1051,45 @@ def hexlify_file(file_path: str):
     with open(file_path, "rb") as f:
         return binascii.hexlify(f.read()).decode("utf-8")
 
+def convert_geojson_to_others_ogr2ogr(
+    input_geojson: dict, destination_layer_name: str, output_format: str
+):
+    options = {
+        "geopackage": {"output_suffix": "gpkg", "format_name": "GPKG"},
+        "shapefile": {"output_suffix": "shp", "format_name": "ESRI Shapefile"},
+        "csv": {"output_suffix": "csv", "format_name": "CSV"},
+        "kml": {
+            "output_suffix": "kml",
+            "format_name": "KML",
+            "extra_options": "-mapFieldType Integer64=Real",
+        },
+        "geobuf": {"output_suffix": "fgb", "format_name": "FlatGeobuf"},
+        "xlsx": {"output_suffix": "xlsx", "format_name": "XLSX"},
+    }
+    
+    output_suffix = options[output_format]["output_suffix"]
+    format_name = options[output_format]["format_name"]
+    extra_options = options[output_format].get("extra_options", "")
+
+    temp_file_base_name = tempfile.mktemp()
+    geojson_temp_file = temp_file_base_name + ".geojson"
+    output_temp_file = temp_file_base_name + f".{output_suffix}"
+
+    with open(geojson_temp_file, "w") as f:
+        f.write(json.dumps(input_geojson))
+
+    command_to_convert = f"ogr2ogr -f {format_name} -nln {destination_layer_name} {extra_options} {output_temp_file} {geojson_temp_file}"
+    subprocess.check_output(
+        command_to_convert, shell=True
+    )  # Use check output to raise error if command fails
+    delete_file(geojson_temp_file)
+    output_data = open(output_temp_file, "rb").read()
+    delete_file(output_temp_file)
+    
+    return {
+        "data": output_data,
+        "output_suffix": output_suffix,
+    }
 
 def read_results(results, return_type=None):
     """
